deployment: dramatically simplify slack messages
