+++ enterprise/dev/deployment-lag-notifer/main.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"flag"
 	"fmt"
 	"io"
@@ -11,8 +9,6 @@ import (
 	"net/http"
 	"os"
 	"strings"
-	"text/tabwriter"
-	"text/template"
 	"time"
 
 	"github.com/sourcegraph/sourcegraph/lib/errors"
@@ -91,139 +87,6 @@ func getLiveVersion(client *http.Client, url string) (string, error) {
 	return version, nil
 }
 
-// GithubResponse is the response payload from requesting GET /repos/:author/:repo/commits
-type GithubResponse []struct {
-	Sha    string `json:"sha"`
-	Commit struct {
-		Author struct {
-			Name string `json:"name"`
-		} `json:"author"`
-		Message string `json:"message"`
-	} `json:"commit"`
-}
-
-// Commit is a singular Git commit to a repo
-type Commit struct {
-	Sha     string
-	Author  string
-	Message string
-}
-
-func getCommitLog(client *http.Client) ([]Commit, error) {
-	var commits []Commit
-
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	url := "https://api.github.com/repos/sourcegraph/sourcegraph/commits"
-
-	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
-	if err != nil {
-		return commits, err
-	}
-
-	q := req.URL.Query()
-	q.Add("branch", "main")
-	q.Add("per_page", "20")
-	q.Add("page", "1")
-
-	req.URL.RawQuery = q.Encode()
-
-	resp, err := client.Do(req)
-	if err != nil {
-		return commits, err
-	}
-
-	if resp.StatusCode > http.StatusOK {
-		return commits, errors.Newf("received non-200 status code %v", resp.StatusCode)
-	}
-
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		return commits, err
-	}
-
-	// fmt.Println(string(body))
-
-	var gh GithubResponse
-	err = json.Unmarshal(body, &gh)
-	if err != nil {
-		return commits, err
-	}
-
-	for _, g := range gh {
-		lines := strings.Split(g.Commit.Message, "\n")
-		message := g.Sha[:7]
-		commits = append(commits,
-			Commit{Sha: message, Author: g.Commit.Author.Name, Message: lines[0]})
-	}
-
-	return commits, nil
-}
-
-// checkForCommit checks for the current version in the
-// last 20 commits
-func checkForCommit(version string, commits []Commit) bool {
-	found := false
-	for _, c := range commits {
-		if c.Sha == version[:7] {
-			found = true
-		}
-	}
-
-	return found
-}
-
-// createMessage posts the message to Slack or stdout if dry is set
-func createMessage(version, environment string, commits []Commit) (bytes.Buffer, error) {
-	var msg bytes.Buffer
-	var slackTemplate = `:worried: *{{.Environment}}*'s deployments are out of date
-
-Current version: ` + "`{{ .Version }}`" + `
-
-Last 20 commits:
-` + "```" + `
-SHA{{"\t"}}Author{{"\t"}}Message
----{{"\t"}}---{{"\t"}}---
-{{- range .CommitLog }}
-{{ .Sha }}{{"\t"}}{{ .Author }}{{"\t"}}{{ .Message }}
-{{- end }}
-` + "```" + `
-
-<https://sourcegraph.com|Handbook Page> | <https://github.com/sourcegraph/deploy-sourcegraph-cloud/pulls|deploy-sourcegraph-cloud> 
-
-cc <!subteam^S02NFV6A536|devops-support>
-`
-
-	type templateData struct {
-		Version     string
-		CommitLog   []Commit
-		Environment string
-	}
-
-	td := templateData{Version: version, CommitLog: commits, Environment: environment}
-	// td := templateData{Version: version, Environment: environment}
-
-	tpl, err := template.New("slack-message").Parse(slackTemplate)
-	if err != nil {
-		return msg, err
-	}
-
-	// tw := tabwriter.NewWriter(&msg, 0, 8, 2, '\t', 0)
-	tw := tabwriter.NewWriter(&msg, 0, 8, 1, '\t', 0)
-
-	err = tpl.Execute(tw, td)
-	if err != nil {
-		return msg, err
-	}
-
-	tw.Flush()
-
-	return msg, nil
-}
-
 func main() {
 	flags := &Flags{}
 	flags.Parse()
@@ -240,11 +103,18 @@ func main() {
 		log.Fatal(err)
 	}
 
+	currentCommit, err := getCommit(&client, version)
+	if err != nil {
+		log.Fatal(err)
+	}
+
+	fmt.Println(currentCommit)
+
 	slack := NewSlackClient(flags.SlackWebhookURL)
 
 	current := checkForCommit(version, commitLog)
 	if current || !current {
-		msg, err := createMessage(version[:7], flags.Environment, commitLog)
+		msg, err := createMessage(version[:7], flags.Environment, currentCommit)
 		if !flags.DryRun {
 			err = slack.PostMessage(msg)
 			if err != nil {
