+++ enterprise/dev/deployment-lag-notifer/slack.go
@@ -7,6 +7,8 @@ import (
 	"fmt"
 	"io/ioutil"
 	"net/http"
+	"text/tabwriter"
+	"text/template"
 	"time"
 )
 
@@ -64,4 +66,46 @@ func (s *SlackClient) PostMessage(b bytes.Buffer) error {
 	fmt.Println(string(body))
 
 	return nil
+}
+
+// createMessage posts the message to Slack or stdout if dry is set
+func createMessage(version, environment string, current Commit) (bytes.Buffer, error) {
+	var msg bytes.Buffer
+
+	drift := time.Now().Sub(current.Date).Hours()
+
+	var slackTemplate = `:warning: *{{.Environment}}*'s version was not found in the last 20 commits.
+
+Current version: ` + "`{{ .Version }}`" + ` was committed *{{.Drift}} hours ago*. 
+
+Check <https://github.com/sourcegraph/deploy-sourcegraph-cloud/pulls|deploy-sourcegraph-cloud> to see if a release is blocked.
+
+cc <!subteam^S02NFV6A536|devops-support>
+`
+
+	type templateData struct {
+		Version     string
+		Environment string
+		Drift       string
+	}
+
+	td := templateData{Version: version, Environment: environment, Drift: fmt.Sprintf("%.2f", drift)}
+	// td := templateData{Version: version, Environment: environment}
+
+	tpl, err := template.New("slack-message").Parse(slackTemplate)
+	if err != nil {
+		return msg, err
+	}
+
+	// tw := tabwriter.NewWriter(&msg, 0, 8, 2, '\t', 0)
+	tw := tabwriter.NewWriter(&msg, 0, 8, 1, '\t', 0)
+
+	err = tpl.Execute(tw, td)
+	if err != nil {
+		return msg, err
+	}
+
+	tw.Flush()
+
+	return msg, nil
 }
